Move extension methods to more discoverable namespaces
Fixes #112
