+++ samples/MvcApp/Global.asax.cs
@@ -1,8 +1,8 @@
+using System.Web;
 using System.Web.Http;
 using System.Web.Mvc;
 using System.Web.Optimization;
 using System.Web.Routing;
-using Microsoft.AspNetCore.SystemWebAdapters;
 
 namespace MvcApp
 {
