+++ samples/RemoteAuth/Bearer/RemoteBearer/Global.asax.cs
@@ -1,6 +1,6 @@
+using System.Web;
 using System.Web.Http;
 using System.Web.Mvc;
-using Microsoft.AspNetCore.SystemWebAdapters;
 
 namespace RemoteOAuth
 {
