+++ src/Microsoft.AspNetCore.SystemWebAdapters/Authentication/RemoteAppAuthenticationExtensions.Framework.cs
@@ -1,12 +1,11 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System;
-using System.Web;
+using Microsoft.AspNetCore.SystemWebAdapters;
 using Microsoft.AspNetCore.SystemWebAdapters.Authentication;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Microsoft.AspNetCore.SystemWebAdapters;
+namespace System.Web;
 
 /// <summary>
 /// Helper methods for registering remote authentication endpoints.
